+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/PostgresLedgerExtenderService.cs
@@ -736,12 +736,10 @@ UPDATE pending_transactions
         var vaultChanges = new List<IVaultChange>();
         var nonFungibleIdChanges = new List<NonFungibleIdChange>();
         var resourceSupplyChanges = new List<ResourceSupplyChange>();
-        var validatorSetChanges = new List<ValidatorSetChange>();
         var vaultHistoryToAdd = new List<EntityVaultHistory>();
         var schemaHistoryToAdd = new List<SchemaHistory>();
         var nonFungibleSchemaHistoryToAdd = new List<NonFungibleSchemaHistory>();
         var keyValueStoreSchemaHistoryToAdd = new List<KeyValueStoreSchemaHistory>();
-        var validatorKeyHistoryToAdd = new Dictionary<ValidatorKeyLookup, ValidatorPublicKeyHistory>();
         var accountDefaultDepositRuleHistoryToAdd = new List<AccountDefaultDepositRuleHistory>();
         var accountResourcePreferenceRuleHistoryToAdd = new List<AccountResourcePreferenceRuleHistory>();
         var validatorEmissionStatisticsToAdd = new List<ValidatorEmissionStatistics>();
@@ -754,6 +752,7 @@ UPDATE pending_transactions
         var packageCodeProcessor = new PackageCodeProcessor(processorContext, networkConfiguration.Id);
         var packageBlueprintProcessor = new PackageBlueprintProcessor(processorContext, referencedEntities);
         var keyValueStoreProcessor = new KeyValueStoreProcessor(processorContext);
+        var validatorProcessor = new ValidatorProcessor(processorContext, referencedEntities);
 
         // step: scan all substates & events to figure out changes
         {
@@ -843,21 +842,6 @@ UPDATE pending_transactions
                                 stateVersion));
                         }
 
-                        if (substateData is CoreModel.ValidatorFieldStateSubstate validator)
-                        {
-                            var lookup = new ValidatorKeyLookup(referencedEntities.Get((EntityAddress)substateId.EntityAddress).DatabaseId, validator.Value.PublicKey.KeyType.ToModel(),
-                                validator.Value.PublicKey.GetKeyBytes());
-
-                            validatorKeyHistoryToAdd[lookup] = new ValidatorPublicKeyHistory
-                            {
-                                Id = sequences.ValidatorPublicKeyHistorySequence++,
-                                FromStateVersion = stateVersion,
-                                ValidatorEntityId = lookup.ValidatorEntityId,
-                                KeyType = lookup.PublicKeyType,
-                                Key = lookup.PublicKey,
-                            };
-                        }
-
                         if (substateData is CoreModel.ConsensusManagerFieldStateSubstate consensusManagerFieldStateSubstate)
                         {
                             if (consensusManagerFieldStateSubstate.Value.Round == 0)
@@ -865,21 +849,6 @@ UPDATE pending_transactions
                                 newEpoch = consensusManagerFieldStateSubstate.Value.Epoch;
                                 passingEpoch = newEpoch - 1;
                             }
-                        }
-
-                        if (substateData is CoreModel.ConsensusManagerFieldCurrentValidatorSetSubstate validatorSet)
-                        {
-                            var change = validatorSet.Value.ValidatorSet
-                                .ToDictionary(
-                                    v =>
-                                    {
-                                        var vid = referencedEntities.Get((EntityAddress)v.Address).DatabaseId;
-
-                                        return new ValidatorKeyLookup(vid, v.Key.KeyType.ToModel(), v.Key.GetKeyBytes());
-                                    },
-                                    v => TokenAmount.FromDecimalString(v.Stake));
-
-                            validatorSetChanges.Add(new ValidatorSetChange(passingEpoch!.Value, change, stateVersion));
                         }
 
                         if (substateData is CoreModel.AccountFieldStateSubstate accountFieldState)
@@ -974,6 +943,7 @@ UPDATE pending_transactions
                         packageCodeProcessor.VisitUpsert(substateData, referencedEntity, stateVersion);
                         packageBlueprintProcessor.VisitUpsert(substateData, referencedEntity, stateVersion);
                         keyValueStoreProcessor.VisitUpsert(substateData, referencedEntity, stateVersion);
+                        validatorProcessor.VisitUpsert(substateData, referencedEntity, stateVersion, passingEpoch);
                     }
 
                     foreach (var deletedSubstate in stateUpdates.DeletedSubstates)
@@ -1201,13 +1171,13 @@ UPDATE pending_transactions
             var mostRecentResourceEntitySupplyHistory = await readHelper.MostRecentResourceEntitySupplyHistoryFor(resourceSupplyChanges, token);
             var mostRecentEntityNonFungibleVaultHistory = await readHelper.MostRecentEntityNonFungibleVaultHistory(vaultSnapshots.OfType<NonFungibleVaultSnapshot>().ToList(), token);
             var existingNonFungibleIdData = await readHelper.ExistingNonFungibleIdDataFor(nonFungibleIdChanges, vaultSnapshots.OfType<NonFungibleVaultSnapshot>().ToList(), token);
-            var existingValidatorKeys = await readHelper.ExistingValidatorKeysFor(validatorSetChanges, token);
 
-            await entityMetadataProcessor.LoadMostRecent();
-            await componentMethodRoyaltyProcessor.LoadMostRecent();
-            await entityRoleAssignmentProcessor.LoadMostRecent();
-            await packageCodeProcessor.LoadMostRecent();
-            await packageBlueprintProcessor.LoadMostRecent();
+            await entityMetadataProcessor.LoadDependencies();
+            await componentMethodRoyaltyProcessor.LoadDependencies();
+            await entityRoleAssignmentProcessor.LoadDependencies();
+            await packageCodeProcessor.LoadDependencies();
+            await packageBlueprintProcessor.LoadDependencies();
+            await validatorProcessor.LoadDependencies();
             await keyValueStoreProcessor.LoadMostRecent();
 
             dbReadDuration += sw.Elapsed;
@@ -1226,6 +1196,7 @@ UPDATE pending_transactions
             packageCodeProcessor.ProcessChanges();
             packageBlueprintProcessor.ProcessChanges();
             keyValueStoreProcessor.ProcessChanges();
+            validatorProcessor.ProcessChanges();
 
             foreach (var e in nonFungibleIdChanges)
             {
@@ -1541,20 +1512,6 @@ UPDATE pending_transactions
                 })
                 .ToList();
 
-            var validatorActiveSetHistoryToAdd = validatorSetChanges
-                .SelectMany(e =>
-                {
-                    return e.ValidatorSet.Select(vs => new ValidatorActiveSetHistory
-                    {
-                        Id = sequences.ValidatorActiveSetHistorySequence++,
-                        FromStateVersion = e.StateVersion,
-                        Epoch = e.Epoch,
-                        ValidatorPublicKeyHistoryId = existingValidatorKeys.GetOrAdd(vs.Key, _ => validatorKeyHistoryToAdd[vs.Key]).Id,
-                        Stake = vs.Value,
-                    });
-                })
-                .ToList();
-
             var entityResourceAggregateHistoryToAdd = entityResourceAggregateHistoryCandidates.Where(x => x.ShouldBePersisted()).ToList();
             var entityResourceVaultAggregateHistoryToAdd = entityResourceVaultAggregateHistoryCandidates.Where(x => x.ShouldBePersisted()).ToList();
 
@@ -1574,8 +1531,6 @@ UPDATE pending_transactions
             rowsInserted += await writeHelper.CopyNonFungibleIdStoreHistory(nonFungibleIdStoreHistoryToAdd.Values, token);
             rowsInserted += await writeHelper.CopyNonFungibleIdLocationHistory(nonFungibleIdLocationHistoryToAdd, token);
             rowsInserted += await writeHelper.CopyResourceEntitySupplyHistory(resourceEntitySupplyHistoryToAdd, token);
-            rowsInserted += await writeHelper.CopyValidatorKeyHistory(validatorKeyHistoryToAdd.Values, token);
-            rowsInserted += await writeHelper.CopyValidatorActiveSetHistory(validatorActiveSetHistoryToAdd, token);
             rowsInserted += await writeHelper.CopySchemaHistory(schemaHistoryToAdd, token);
             rowsInserted += await writeHelper.CopyAccountDefaultDepositRuleHistory(accountDefaultDepositRuleHistoryToAdd, token);
             rowsInserted += await writeHelper.CopyAccountResourcePreferenceRuleHistory(accountResourcePreferenceRuleHistoryToAdd, token);
@@ -1590,6 +1545,7 @@ UPDATE pending_transactions
             rowsInserted += await packageCodeProcessor.SaveEntities();
             rowsInserted += await packageBlueprintProcessor.SaveEntities();
             rowsInserted += await keyValueStoreProcessor.SaveEntities();
+            rowsInserted += await validatorProcessor.SaveEntities();
 
             await writeHelper.UpdateSequences(sequences, token);